+++ ProtronServer/src/main/java/com/Protronserver/Protronserver/Entities/TimesheetTask.java
@@ -16,7 +16,7 @@ public class TimesheetTask {
     private Long taskId;
     private String taskType;
     private Date date;
-    private Integer hoursSpent;
+    private Double hoursSpent;
     private String description;
     private boolean isSubmitted = false;
 
@@ -64,11 +64,11 @@ public class TimesheetTask {
         this.date = date;
     }
 
-    public Integer getHoursSpent() {
+    public Double getHoursSpent() {
         return hoursSpent;
     }
 
-    public void setHoursSpent(Integer hoursSpent) {
+    public void setHoursSpent(Double hoursSpent) {
         this.hoursSpent = hoursSpent;
     }
